# gateway_service/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Единая точка входа для всех микросервисов медицинской платформы Vitalem. Упрощает взаимодействие с backend сервисами через простые и понятные REST API.",
        "title": "Vitalem API Gateway",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8800",
    "basePath": "/",
    "paths": {
        "/appointments": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получение списка записей текущего пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Записи"
                ],
                "summary": "Мои записи",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Количество записей",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список записей",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "appointments": {
                                    "type": "array"
                                },
                                "total": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/appointments/exceptions": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получение списка исключений в расписании врача за период",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Исключения в расписании"
                ],
                "summary": "Исключения врача",
                "parameters": [
                    {
                        "type": "string",
                        "example": "\"2024-06-01\"",
                        "description": "Дата начала в формате YYYY-MM-DD",
                        "name": "start_date",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "example": "\"2024-06-30\"",
                        "description": "Дата окончания в формате YYYY-MM-DD",
                        "name": "end_date",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список исключений",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "exceptions": {
                                    "type": "array"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверные параметры",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Только врачи могут просматривать исключения",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Добавление исключения в расписание врача (выходной день или изменение часов)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Исключения в расписании"
                ],
                "summary": "Создать исключение в расписании",
                "parameters": [
                    {
                        "description": "Данные исключения",
                        "name": "exceptionData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "custom_end_time": {
                                    "type": "string"
                                },
                                "custom_start_time": {
                                    "type": "string"
                                },
                                "date": {
                                    "type": "string"
                                },
                                "reason": {
                                    "type": "string"
                                },
                                "type": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Исключение создано",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "date": {
                                    "type": "string"
                                },
                                "id": {
                                    "type": "string"
                                },
                                "reason": {
                                    "type": "string"
                                },
                                "type": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Только врачи могут создавать исключения",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/appointments/schedules": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получение всех расписаний текущего врача",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Расписание"
                ],
                "summary": "Мои расписания",
                "responses": {
                    "200": {
                        "description": "Список расписаний",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "schedules": {
                                    "type": "array"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Только врачи могут просматривать расписания",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создание нового расписания работы врача",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Расписание"
                ],
                "summary": "Создать расписание",
                "parameters": [
                    {
                        "description": "Данные расписания",
                        "name": "scheduleData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "end_time": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "slot_duration": {
                                    "type": "integer"
                                },
                                "start_time": {
                                    "type": "string"
                                },
                                "work_days": {
                                    "type": "array"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Расписание создано",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "name": {
                                    "type": "string"
                                },
                                "work_days": {
                                    "type": "array"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Только врачи могут создавать расписания",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/doctors": {
            "get": {
                "description": "Получение списка всех активных врачей (публичный доступ)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Врачи"
                ],
                "summary": "Список врачей",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "Количество записей",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Смещение",
                        "name": "offset",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Фильтр по специализации",
                        "name": "specialization",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Список врачей",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "doctors": {
                                    "type": "array"
                                },
                                "total": {
                                    "type": "integer"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Создание нового профиля врача (только для авторизованных пользователей)",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Врачи"
                ],
                "summary": "Создать профиль врача",
                "parameters": [
                    {
                        "description": "Данные врача",
                        "name": "doctorData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "first_name": {
                                    "type": "string"
                                },
                                "last_name": {
                                    "type": "string"
                                },
                                "specializations": {
                                    "type": "array"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Профиль создан",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Неверные данные",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/notifications/my": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получение списка уведомлений текущего пользователя",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Уведомления"
                ],
                "summary": "Мои уведомления",
                "responses": {
                    "200": {
                        "description": "Список уведомлений",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "notifications": {
                                    "type": "array"
                                },
                                "total_count": {
                                    "type": "integer"
                                },
                                "user_id": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Неавторизован",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Внутренняя ошибка",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите JWT токен в формате: Bearer {ваш_токен}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}